only delete temp rprof files when hosted in the profiles-cache directory

+++ src/cpp/session/modules/SessionProfiler.R
@@ -107,7 +107,10 @@
          profvis <- profilerOptions$profvis
          htmlwidgets::saveWidget(profvis, htmlFile, selfcontained = TRUE)
 
-         file.remove(profilerOptions$profvis$x$message$prof_output)
+         if (resources$tempPath == substr(profilerOptions$profvis$x$message$prof_output, 1, nchar(resources$tempPath)))
+         {
+            file.remove(profilerOptions$profvis$x$message$prof_output)
+         }
       }
 
       return(list(
